requirements.txt
# PyTorch with CUDA 12.8 support
torch==2.7.1+cu128
torchvision==0.18.1+cu128
torchaudio==2.7.1+cu128

# Core dependencies
transformers>=4.20.0
Pillow>=8.0.0
numpy>=1.21,<1.25
tqdm>=4.60.0

# OpenCLIP for ViT-g-14 model
open_clip_torch>=2.24.0

# Environment and utilities
python-dotenv>=0.19.0

# Optional: GPU-accelerated similarity search
# conda install -c conda-forge faiss-gpu=1.7.4

# Installation notes:
# 1. PyTorch packages with CUDA 12.8 support:
#    pip install torch==2.7.1+cu128 torchvision==0.18.1+cu128 torchaudio==2.7.1+cu128 --index-url https://download.pytorch.org/whl/cu128
# 2. OpenCLIP installation:
#    pip install open_clip_torch
# 3. For CUDA 12.8 compatibility, ensure NVIDIA driver version >= 573.22